added handling exceptions for inputs

+++ src/main/java/operations/Addition.java
@@ -5,17 +5,15 @@ package operations;
  */
 public class Addition extends Operation {
 
-    public Addition(float n1, float n2) {
+    public Addition(double n1, double n2) {
         super(n1, n2);
     }
 
-    public float calculateResult() { //performing the addition
+    public double calculateResult() { //performing the addition
         return getN1() + getN2();
     }
 
     public String getOperationSign(){ //getting the actual operation sign for the final output
         return "+";
     }
-
-
 }
